Restructure crate to contain lib and bin
The binary name will be the same. The binary can be built like

```
cargo build --bin fitbit-grabber
```

The new library name is "fitbit", and src/bin/main.rs imports it. We could
configure to have more binaries, but only one top-level library.

Refs #1
