+++ .gitignore
@@ -1,4 +1,8 @@
 /target/
 **/*.rs.bk
 
+# Editors and IDEs
+.idea/
+
+# token file
 .token
